+++ src/onegov.form/onegov/form/parser/form.py
@@ -17,10 +17,14 @@ from wtforms import StringField
 from wtforms import TextAreaField
 from wtforms.fields.html5 import DateField
 from wtforms.fields.html5 import DateTimeLocalField
+from wtforms.fields.html5 import DecimalField
 from wtforms.fields.html5 import EmailField
 from wtforms.fields.html5 import IntegerField
-from wtforms.fields.html5 import DecimalField
-from wtforms.validators import DataRequired, Length, Optional, NumberRange
+from wtforms.validators import DataRequired
+from wtforms.validators import Length
+from wtforms.validators import NumberRange
+from wtforms.validators import Optional
+from wtforms.validators import Regexp
 from wtforms.widgets import TextArea
 from wtforms_components import Email, If, TimeField
 
@@ -59,10 +63,14 @@ def handle_field(builder, field, dependency=None):
     """ Takes the given parsed field and adds it to the form. """
 
     if field.type == 'text':
+
         if field.maxlength:
             validators = [Length(max=field.maxlength)]
         else:
             validators = []
+
+        if field.regex:
+            validators.append(Regexp(field.regex))
 
         builder.add_field(
             field_class=StringField,